徐虹璐：加入了课程表和ai分析

+++ src/main/java/com/bupt/tarecruitment/admin/AcceptedAssignment.java
@@ -10,7 +10,7 @@ public record AcceptedAssignment(
     String jobId,
     String title,
     String moduleOrActivity,
-    int weeklyHours,
+    double weeklyHours,
     List<String> scheduleSlots
 ) {
     public AcceptedAssignment {
